Refact(CarInfo) : CarInfo 객체 리팩토링
